binance/requirements.txt
aio-pika>=8.0.3
python-dotenv>=0.20.0
uvloop>=0.16.0
websockets>=10.3
